feat(movement): allow movement in all directions
This commit extends OperationDirection to include the four directions in
addition to Previous/Next.

The Swap command in yattac is replaced by the Move command, and a bunch
of scattered logic in the socket message handler has been consolidated
in the DirectionOperation enum which has a handler to decide how to
respond to Move and Focus ops.