fix(tiling): allow recognition and tiling of uwp apps
Borrowed and repurposed some more code from Nog and learnt about bitmaps
along the way. should_listen is replaced by a modified version of Nog's
should_manage along with the bitmap structs that fn depends on.

Looks like windows-rs is missing some stuff around cloaked window
detection for UWP apps, but I found some code from screenshot-rs that
handles this nicely which I also repurposed for use here. I can now
finally tile the Calendar app!

I was getting spammed by ObjectNameChange events with the hack to
compensate for Firefox not sending a proper Create or Show event when it
is opened for the first time, so I restricted this to only allow
messages through for Firefox. Still not great since the messages will go
through every time a tab changes, but the dedupe logic in the message
handler avoids any needless redrawing of the workspace for now.