feat(tiling): make gaps configurable
Not sure how to remove gaps completely yet, but this commit makes them
configurable to a specific value or via incrementing and decrementing
using yattac. Might do the same with padding later, but the default
padding is fine for me right now.

Decided to add the Clap dependency to yatta-core after all because
having duplicated enums and structs between yatta-core and yattac was
starting to bug me. Cleaned up the dupe structs in yattac now that the
Clap derive blocks can go into yatta-core.