feat(tiling): add float toggle
This commit allows the user to toggle the focused window between tiling
and floating mode. Ended up expanding the Window struct to hold the
state for whether or not it should be tiled and thank the gods for
IntelliJ or I would have gone nuts changing the usages everywhere by
hand.