feat(tiling): enable toggling between horizontal and vertical orientations
This commit adds a toggle and a set command for the tiling orientation.
Since I'm still using a relatively simple Vec to represent the tree,
I've had to switch on the window_op_* functions to behave differently
depending on the orientation of the layout.

Cleaned up some repetition in yattac that was getting a little out of
control.

Clap gives weird panics trying to unwrap Options when a command defined
in the Enum to take an arg doesn't have an arg supplied. Not going to
bother looking into that seriously; it works when an arg is passed which
is good enough for me.