feat(tiling)!: add columns, rows and monocle as layouts
This commit adds a Layouts enum and adds Columns, Rows and Monocle in
addition to deprecating Orientation and adding BSPV and BSPH as Layout
enum variants. Layouts can be set specifically using yattac or cycled
through.

Once the final Layout variant has been reached, trying to cycle to the
next variant will take the user back to the start, and the behaviour is
the same when cycling in reverse.

BREAKING CHANGE: Orientation deprecated now that there are more types of
layouts in addition to BSP. The Horizontal and Vertical Orientation enum
variants now live in the Layout enum as BSPH and BSHV respectively. As a
result, any yattac commands that used to operate on Orientations no
longer exist.