feat(cli): add start and stop commands
It looks like setting up a Windows Service is a fairly involved process
that will require more code. In the meantime, I have added start and
stop commands to yattac that shell out Powershell scripts to launch and
kill the process in the background so that people can have this running
with a binary instead of needing to do 'cargo run' on a checked out
repo.