feat(tiling): add per class and exe floating rules
This commit allows for the setting of rules on either class or exe names
which will determine whether a window should always be floated.

Added some examples in the readme with IntelliJ popups, the Control
Panel window and Wally.exe.

Also fixed a bug where the Columns and Rows layouts would not take into
account floated windows when calculating their layouts.