feat(tiling): add floating rules based on partial title matching
This commit allows for the setting of floating rules based on a partial
title match.

Readme has been updated with an example added to yatta.ahk
showing a title based floating rule for the Calculator app.

Added a check in yatta to ensure that when running yattac.exe's start
command, duplicate processes are not created (the second proc exits with
code 1 immediately).