feat(tiling): add rules to assign exes to desktops
This commit adds new configuration options that allow us to ensure a
certain number of virtual desktops exist, and to assign certain
executables to specific desktops. Some general tidying up and renaming.

Tried playing around a little with winvd to send windows to different
desktops and switch desktops, but the results were very unreliable when
calling the commands through AHK. Too much going on in the event stream
in different places forcing too many redraws and not staying on the
correct desktop with the current implementation. Might need to take a
look at using winit to replace all of this current window spaghetti
code.