refactor(logging): use flexi_logger
This commit swaps out env_logger for flexi_logger and generally tidies
up the logging. Logs now go to yatta.log at DEBUG and to the console at
INFO.

Changed a bunch of unwraps to IFLS and IFLOK statements, and added
an explicit HWND checker/cleaner at the start of every message handled
on the yatta message channel to purge any windows that no longer exist.

Finally figured out how to ensure focus is taken whenever using the
yattac focus command, but it requires a bit of a hack in AHK to send an
Alt key before running the yattac command. I initially tried doing this
in code, and while it worked technically, the release of the Alt key in
code made it impossible to hold Alt and press the navigation keys to
switch window focus multiple times.