revert(tiling)!: remove virtual desktop commands
After some more reading I think it was a mistake to add any commands
that hook into Windows Virtual Desktops given the instability of that
API and the fact that it does't emit events in the normal Windows event
stream, or expose any methods to gather the state of windows across all
workspaces.

If I decide to do something with workspaces, it will have to be a thin
layer on top of a single Virtual Desktop that hides/shows windows as a
user switches workspaces.

BREAKING CHANGE: Removed all commands that deal with Virtual Desktops
via the winvd crate