feat(tiling): enumerate through displays and look up display handles for all windows
This commit adds the tracking of multiple monitors to a Desktop. For now
at least, a Desktop is comprised of one or more monitors since if I ever
decide to implement virtual desktops in the future, a desktop will span
across multiple monitors.

The dimensions of a monitor get set on the Display struct, and in a
future commit this is what will be referenced to calculate the
geometries of windows on different monitors. The Window struct has been
updated to include the hmonitor reference. Next step will be to make the
geometry calculation code run across monitors while taking the hmonitor
reference on each window into consideration.