feat(tiling): push window and layout state down to the display level
With this commit, the management of Window and Layout state, including
all operations, gets pushed from the Desktop struct to the nested
Display structs.

This should, in theory, allow for each Display to keep track of and
manage Windows it knows about. At this point, there is no handling of
Windows that are manually dragged or transferred to other displays. I
imagine they will be snapped back to the display that they were detected
on when the yatta proc started, but that's just a best guess; I guess
that is undefined behaviour at the moment.

Removed some (what I think are) extraneous calls to
get_visible_windows() and everything seems to be working alright, need
to keep testing with this build though to see if this causes any
regressions.