feat(tiling): add commands to move and focus between displays
This commit exposes two new groups of SocketMessages in yatta-core to
move windows between displays using either a target or a CycleDirection,
and also to change focus between displays using either a target or a
CycleDirection. CLI commands have been added to send these messages and
yattac and handlers have been added for them in yatta.

Implementation is pretty much as expected, with any cross-Display state
changes being managed at the Desktop level. The only thing of note is
that when trying to switch focus to a Display which may not have any
windows, the fallback behaviour is to centre the mouse in the target
display and left click to gain focus.

The docs for bindings::windows::win32::keyboard_and_mouse_input::INPUT
seemed a bit thin and I didn't want to figure out how to call
bindings::windows::win32::keyboard_and_mouse_input::SendInput, so I
ended up pulling in the enigo crate that already has nice wrappers
around mouse input events for Windows. Unfortunately this also means
pulling in winapi-rs as a dependency, but I guess that's what I get for
jumping on the windows-rs bandwagon too early.