feat(tiling): add resizing operations for bspv and bsph layouts
This commit exposes the Resize operation via yattac, which takes a
Window Edge and a Sizing.

Adjustments made to the Right and Bottom edges are applied on the Window
the operation is called on, whereas adjustments made to the Left and Top
edges are applied to neighbouring windows at those edges via
calculate_resize_adjustment().

The amount that an edge is increased or decreased by is fixed at the
moment, but since it is a member of the Display struct it could be
adjusted in the future in the same way that the gaps currently are.

Some edge cases that needed to be handled were:

* Trading resize offsets between windows when they are swapped with the
  Move operation
* Removing all resize offsets when retiling to rebalance the current
  layout
* Removing all resize offsets when cycling between or switching layouts
  to ensure every layout loads in a balanced state.