refactor(windows-rs): upgrade to 0.9.1 and handle all breaking changes
This was probably the least fun upgrade I've ever done of a dependency.
Many imports moved packages, contants were changed to associated
constants, and the linker refused to work with wildcards in the build.rs
file of the bindings package.

This commit is the result of patiently going through each error reported
by rustc, cross referencing with the new windows-rs docs and updating /
making things more semantic where it seemed like the right thing to do.

I am running this commit locally for a few days to see if any
regressions have been introduced, but it seems like everything is
alright at first glance.