feat(tiling): add acceptable layered window whitelist
As highlighted in #19, certain applications such as Steam may have
GwlExStyle::LAYERED set when events are emitted by the window being
closed by the user.

This commit introduces a global whitelist that is checked when
evaluating whether or not yatta should try to handle an emitted event in
Window::should_manage().

Some general refactoring of the logic in that fn to add some early
returns, and some comments describing the boolean cases that we are
trying to match on for the desired behaviours.

re #19