feat(tiling): make padding configurable
This commit makes the padding around the edges of the desktop, which was
previously a hard-coded constant, a value that can be configured using
yattac.

Previously this value was only being read once in the
enum_display_monitor callback, but now Desktop.dimensions has been made
a private struct member, and any time the Rect dimensions for a Display
are required, they should be accessed using Display::get_dimensions
which will apply whatever the latest value of PADDING is to the
dimensions returned.

When setting the padding around the edges of the desktop to 0, the
windows still do not get arranged right at the edges of the screen, and
in fact, the padding on the vertical axis will be less than the padding
on the horizontal axis.

This isn't something that I personally have the patience to look into,
but hopefully it will bug somebody enough in the future that they'll be
willing to dig into this behaviour and submit a PR to fix it.

closes #19